get rid of the generalized comm interface since non-canopen roboteq support was dropped
