+++ uwrt_mars_rover_drivetrain/uwrt_mars_rover_drivetrain_hw/drivers/roboteq_driver/include/RoboteqController.hpp
@@ -1,10 +1,9 @@
 #pragma once
 
+#include <CanopenInterface.hpp>
 #include <iostream>
 #include <memory>
 
-//#include "CommunicationInterface.hpp"
-#include "CanopenInterface.hpp"
 namespace roboteq {
 
 class RoboteqController {
