+++ uwrt_mars_rover_drivetrain/uwrt_mars_rover_drivetrain_hw/drivers/roboteq_driver/src/RoboteqController.cpp
@@ -1,4 +1,3 @@
-#include <CommunicationInterface.hpp>
 #include <RoboteqController.hpp>
 
 namespace roboteq {
